react-with-flask/api/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.35
pandas==2.2.2
python-dateutil==2.9.0.post0